build: deployment target 변경 14.5 -> 12.1, Scene Delegate 삭제
