+++ FullStackCodingBot/FullStackCodingBot/View/Delegate/AppDelegate.swift
@@ -2,8 +2,15 @@ import UIKit
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
+    
+    var window: UIWindow?
+    
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        let coordinator = SceneCoordinator(window: window!)
+        let storage = ItemStorage()
+        let mainViewModel = MainViewModel(sceneCoordinator: coordinator, storage: storage)
+        let mainScene = Scene.main(mainViewModel)
+        coordinator.transition(to: mainScene, using: .root, animated: false)
         return true
     }
 }
